ratio safe-fast works

+++ script/app/location_moderator.py
@@ -78,7 +78,7 @@ def change_type(G):
 
     for i in range(len(edges)):
         
-        edges[i][3]["danger"] = int(edges[i][3]["danger"])
+        edges[i][3]["danger"] = float(edges[i][3]["danger"])
         edges[i][3]["travel_time"] = float(edges[i][3]["travel_time"])
         try:
             edges[i][3]["ratio"] = float(edges[i][3]["ratio"])
